backend-new/requirements.txt
# FastAPI and dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
click>=8.1.7
dependency-injector>=4.41.0
python-dotenv>=1.0.0

# Database
aiomysql>=0.2.0
sqlalchemy>=2.0.22
alembic>=1.12.0
greenlet>=3.0.1
cryptography>=41.0.0  # Required for MySQL connection security

# JSON Processing
ujson>=5.7.0

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
PyJWT>=2.7.0

# API clients
httpx>=0.25.0
requests>=2.31.0

# Task queue and messaging
celery>=5.3.4
redis>=5.0.1
aio-pika>=9.3.0

# Redis

# Testing
pytest>=7.4.2
pytest-asyncio>=0.21.1

# API wrappers
perplexity-api>=0.1.0

# Cache settings
fastapi-limiter>=0.1.5

# Additional dependencies that might be needed for deployment
aioredis>=2.0.1
